# Fdmi int