--- a/fdmi/plugins/.gitignore
+++ b/fdmi/plugins/.gitignore
@@ -1,0 +1,1 @@
+fdmi_sample_plugin