--- a/fdmi/plugins/Makefile.sub
+++ b/fdmi/plugins/Makefile.sub
@@ -1,0 +1,2 @@
+fdmi_plugins_fdmi_sample_plugin_SOURCES = fdmi/plugins/fdmi_sample_plugin.c
+EXTRA_DIST	+= fdmi/plugins/fdmi_app