--- a/fdmi/plugins/fdmi_plugin_st.sh
+++ b/fdmi/plugins/fdmi_plugin_st.sh
@@ -17,6 +17,10 @@
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
+# Authors:
+#   hua.huang@seagate.com
+#   yuriy.umanets@seagate.com
+#
 
 
 TOPDIR=$(dirname "$0")/../../
@@ -45,13 +49,13 @@ do_some_kv_operations()
 {
 	local rc=0
 
-	for ((i=0; i < 1; i++)) ; do
+	for ((i=0; i<1; i++)) ; do
 		DIX_FID="12345:12345$i"
 		MOTR_PARAM="-l ${lnet_nid}:$SNS_MOTR_CLI_EP  \
 			    -h ${lnet_nid}:$HA_EP -p $PROF_OPT \
 			    -f $M0T1FS_PROC_ID -s "
 
-		$M0_SRC_DIR/utils/m0kv ${MOTR_PARAM}                                       \
+		"$M0_SRC_DIR/utils/m0kv" ${MOTR_PARAM}                                       \
 					index create "$DIX_FID"                            \
 					      put    "$DIX_FID" "somekey" "somevalue"      \
 					      get    "$DIX_FID" "somekey"                  \
@@ -63,6 +67,32 @@ do_some_kv_operations()
 			rc=$?
 			echo "m0kv failed"
 		}
+
+		echo "Now, let's delete 'key2' from this index. The plugin must show the del op coming"
+		"$M0_SRC_DIR/utils/m0kv" ${MOTR_PARAM}                                       \
+					index del    "$DIX_FID" "key2"                     \
+				 || {
+			rc=$?
+			echo "m0kv index del failed"
+		}
+
+		echo "Now, let's get 'key2' from this index again. It should fail."
+		"$M0_SRC_DIR/utils/m0kv" ${MOTR_PARAM}                                       \
+				index get    "$DIX_FID" "key2"                     \
+				 && {
+			rc=22 # EINVAL to indicate the test is failed
+			echo "m0kv index get expected to fail, but did not."
+		}
+		sleep 1
+		echo "Now, let's delete 'key2' from this index again."
+		echo "It should fail, and the plugin must NOT show the del op coming."
+		"$M0_SRC_DIR/utils/m0kv" ${MOTR_PARAM}                                       \
+                                       index del    "$DIX_FID" "key2"                     \
+                    && {
+                    rc=22 # EINVAL to indicate the test is failed
+                    echo "m0kv index del should fail, but did not"
+		}
+
 	done
 	return $rc
 }
@@ -76,12 +106,29 @@ start_fdmi_plugin()
 		    -h ${lnet_nid}:$HA_EP -p $PROF_OPT    \
 		    -f $M0T1FS_PROC_ID                    "
 
-	$M0_SRC_DIR/fdmi/plugins/fdmi_sample_plugin $MOTR_PARAM -g "$FDMI_FILTER_FID" &
+	"$M0_SRC_DIR/fdmi/plugins/fdmi_sample_plugin" $MOTR_PARAM -g $FDMI_FILTER_FID &
 	sleep 5
 
+	# Checking for the pid of the started plugin process
+	local pid=$(pgrep -f "lt-fdmi_sample_plugin")
+	if test "x$pid" = x; then
+		echo "Failed to start FDMI plugin"
+		rc=22 # EINVAL to indidate plugin start is failed
+	fi
 	return $rc
 }
 
+
+stop_fdmi_plugin()
+{
+	local pid=$(pgrep -f "lt-fdmi_sample_plugin")
+	if test "x$pid" != x; then
+		echo "Terminating ${pid}"
+		kill -TERM "${pid}"
+		wait "${pid}"
+	fi
+	return 0
+}
 
 
 motr_fdmi_plugin_test()
@@ -95,29 +142,29 @@ motr_fdmi_plugin_test()
 	echo "MOTR is UP."
 	echo "Motr client config:"
 	echo
-	echo "HA_addr        : ${lnet_nid}:$HA_EP           "
-	echo "Client_addr    : ${lnet_nid}:$SNS_MOTR_CLI_EP "
-	echo "Profile_fid    : $PROF_OPT                    "
-	echo "Process_fid    : $M0T1FS_PROC_ID              "
-	echo "FDMI_plugin_ep : ${lnet_nid}:$FDMI_PLUGIN_EP  "
-	echo "FDMI_FILTER_FID: $FDMI_FILTER_FID             "
+	echo "HA addr        : ${lnet_nid}:$HA_EP           "
+	echo "Client addr    : ${lnet_nid}:$SNS_MOTR_CLI_EP "
+	echo "Profile fid    : $PROF_OPT                    "
+	echo "Process fid    : $M0T1FS_PROC_ID              "
+	echo "FDMI plugin ep : ${lnet_nid}:$FDMI_PLUGIN_EP  "
+	echo "FDMI filter fid: $FDMI_FILTER_FID             "
 	echo
 	echo
 
-	start_fdmi_plugin
+	start_fdmi_plugin && {
+	    do_some_kv_operations || {
+		# Make the rc available for the caller and fail the test
+		# if kv operations fail.
+		rc=$?
+		echo "Test failed with error $rc"
+	    }
+	}
 
-	do_some_kv_operations
-
-#	wait_and_exit
+	# wait_and_exit
 
 	sleep 3
-
-	local pid=$(pgrep -f "lt-fdmi_sample_plugin")
-	echo "Terminating ${pid}"
-	kill -TERM "${pid}"
-	wait "${pid}"
-
-	return 0
+	stop_fdmi_plugin
+	return $rc
 }
 
 main()
@@ -134,7 +181,7 @@ main()
 
 
 	if [[ $rc -eq 0 ]] && ! motr_fdmi_plugin_test ; then
-		echo "Failed: SNS repair failed.."
+		echo "FDMI plugin test failed."
 		rc=1
 	fi