… fdmi/plugins/fdmi_app
… fdmi/plugins/fdmi_sample_plugin.c
… m0t1fs/linux_kernel/st/m0t1fs_common_inc.sh