@mehjoshi 2021-07-29
Following are the perfline results shared by @maximalezhin for #719:
**Main branch throughput**
```   numClients numSamples objectSize Write Errors Write Tput Read Errors Read Tput
1         150      40000      0.250            0     56.972           0   645.236
2         150      40000      0.250            0     69.126           0   962.675
3         150      40000      0.250            0     74.535           0  1146.828
4         300      40000      0.250            0     75.843           0  1152.816
5         300      40000      0.250            0     71.115           0  1010.279
6         300      40000      0.250            0     73.401           1  1104.920
7         450      40000      0.250            0     75.589           0  1055.024
8         450      40000      0.250            0     74.568           0   963.092
9         450      40000      0.250            0     79.979           0  1157.394
10        600      40000      0.250            0     88.401           0   785.309
11        600      40000      0.250            0     88.255           0   717.152
12        600      40000      0.250            0     85.406           0  1140.604
13        150      40000     16.000            0   1207.413           0  4616.082
14        150      40000     16.000            0   1286.534           0  4686.454
15        150      40000     16.000            0   1182.002           0  4707.880
16        300      40000     16.000            0   1255.705           0  4908.331
17        300      40000     16.000            0   1273.240           0  4735.723
18        300      40000     16.000            0   1394.392           0  4824.450
19        450      40000     16.000            0   1228.708           0  4650.537
20        450      40000     16.000            0   1328.574           0  5110.215
21        450      40000     16.000            0   1337.793           0  5173.121
22        600      40000     16.000            0   1429.882           0  5425.798
23        600      40000     16.000            0   1309.105           0  5177.837
24        600      40000     16.000            0   1443.890           0  5416.090
```
**PR #719 throughput:**
```
   numClients numSamples objectSize Write Errors Write Tput Read Errors Read Tput
1         150      40000      0.250            0     62.329           0  1199.825
2         150      40000      0.250            0     67.795           0  1025.906
3         150      40000      0.250            0     80.101           0  1047.009
4         300      40000      0.250            0     82.862           0   851.753
5         300      40000      0.250            0     77.224           0  1086.356
6         300      40000      0.250            0     77.151           0   790.644
7         450      40000      0.250            0     77.771           0   989.379
8         450      40000      0.250            0     79.836           0   880.268
9         450      40000      0.250            0     80.499           0  1267.218
10        600      40000      0.250            0     82.940           0   853.541
11        600      40000      0.250            0     84.297           0   946.789
12        600      40000      0.250            0     81.567           0  1108.828
13        150      40000     16.000            0   1206.375           0  4767.778
14        150      40000     16.000            0   1220.290           0  5102.533
15        150      40000     16.000            0   1185.868           0  4762.672
16        300      40000     16.000            0   1316.799           0  4767.146
17        300      40000     16.000            0   1227.819           0  5250.778
18        300      40000     16.000            0   1315.196           0  5341.022
19        450      40000     16.000            0   1327.333           0  5306.479
20        450      40000     16.000            0   1350.638           0  5260.593
21        450      40000     16.000            0   1312.299           0  5444.710
22        600      40000     16.000            0   1375.142           0  5010.602
23        600      40000     16.000            0   1348.713           0  5256.901
24        600      40000     16.000            0   1311.584           0  5499.078
```